agents/requirements.txt
pyautogen>=0.10.0
langchain>=1.0.0
ollama>=0.6.0


